services/ai-engine/requirements.txt
# Core AI/ML libraries
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
spacy>=3.6.0
nltk>=3.8.0
pandas>=2.0.0
numpy>=1.24.0

# Portuguese language support
spacy[pt_core_news_sm]

# Document processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database and caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
redis>=4.5.0

# Monitoring and telemetry
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-exporter-jaeger>=1.19.0
prometheus-client>=0.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
joblib>=1.3.0
celery>=5.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0